+++ frontend/src/components/MoEntry/MoOrganisationUnitEntry.vue
@@ -13,7 +13,7 @@
         required
       />
 
-      <mo-input-text
+      <mo-input-text v-if="showUserKey"
         :label="$t('input_fields.org_unit_user_key')"
         :placeholder="$t('input_fields.org_unit_user_key_placeholder')"
         v-model="entry.user_key"
@@ -89,6 +89,9 @@ export default {
         return showTimePlanning
       }
       return false
+    },
+    showUserKey () {
+      return !this.isEdit
     }
   },
 
@@ -99,7 +102,7 @@ export default {
     entry: {
       handler (newVal) {
         if (newVal.user_key === undefined || newVal.user_key === ""){
-          newVal.user_key = null; 
+          newVal.user_key = null;
         }
         this.$emit('input', newVal)
       },